// packages/particles2d-core/src/types.ts
export type ParticleConfig = {
    sourcePosition: [number, number];
    sourcePositionSpread: [number, number];
    speed: number;
    speedSpread: number;
    particleLifeSpan: number;
    particleLifespanSpread: number;
    angle: number;
    angleSpread: number;
    gravity: [number, number];
    radialAcceleration: number;
    tangentialAcceleration: number;
    radialAccelSpread: number;
    tangentialAccelSpread: number;
    startColor: [number, number, number, number];
    startColorSpread: [number, number, number, number];
    endColor: [number, number, number, number];
    endColorSpread: [number, number, number, number];
    particlesCount: number;
    startParticleSize: number;
    startParticleSizeSpread: number;
    endParticleSize: number;
    endParticleSizeSpread: number;
    duration: number;
    emitterType: number;
    maxRadius: number;
    maxRadiusSpread: number;
    minRadius: number;
    minRadiusSpread: number;
    rotatePerSecond: number;
    rotatePerSecondSpread: number;
    blendMode: 0 | 1 | 2 | 3;
    rotationStart: number;
    rotationStartSpread: number;
    rotationEnd: number;
    rotationEndSpread: number;
};